renamed header to headers.jsx, added second header for landing page
